requirements.txt
Python==3.8.1
django>=2.2.10
django-filter==2.2.0
djangorestframework==3.11.0